public/python/requirements.txt
flask==1.1.1
pyotp==0.84
pyqrcode==1.2.1
mysql-connector==2.2.9